+++ src/components/shared/templates/StageSubNavigation/index.tsx
@@ -21,12 +21,13 @@
 import { ParentSubNavigation } from '@catena-x/portal-shared-components'
 import '../Templates.scss'
 import { Box } from '@mui/material'
+import { type SubNavigationType } from 'services/CommonService'
 
 export const StageSubNavigation = ({
   linkArray,
   fixHeader,
 }: {
-  linkArray: any
+  linkArray: SubNavigationType[]
   fixHeader: boolean
 }) => {
   const scrollToId = (id: string) => {